Fix for #260: hmmsearch segfault with large HMM vs large protein
The example is a profile of length M=66439 vs. a sequence of length
L=80782, where (ML) = 5.4G cells, overflowing 32-bit calculations in
allocating and reallocating DP matrices in p7_omx_Create() and
p7_omx_GrowTo(). The fix is to add casts to int64_t in the relevant
calculations, including offsetting of DP row pointers.

I've only tested the SSE version as yet. I put exactly the same
changes in the NEON and VMX p7_omx.c code, but I haven't yet built or
tested on ARM or VMX platforms.