Fixes gcc -Walloc-size-larger-than warnings from some ESL_ALLOC calls
On an Ubuntu Linux VM, with gcc 11.3.0 and default ../configure,
a couple of warnings like:

../../easel/easel.h:164:17: warning: argument 1 range [18446744022169944064, 18446744073709551592] exceeds maximum object size 9223372036854775807 [-Walloc-size-larger-than=]
  164 |     if ( ((p) = malloc(size)) == NULL)  { \
      |                 ^~~~~~~~~~~~
../../src/hmmbuild.c:562:3: note: in expansion of macro ‘ESL_ALLOC’
  562 |   ESL_ALLOC(info, sizeof(*info) * (size_t) infocnt);
      |   ^~~~~~~~~

Some versions of gcc have a -Walloc-size-larger-than=PTRDIFF_MAX
option turned on by default, which aggressively checks malloc() calls
for a particular kind of unsigned/signed arithmetic error that arises
on constructs like `malloc(sizeof(foo) * n)`. sizeof() returns size_t,
which is unsigned; so if n is a signed int, it is converted to
unsigned. That means if n was negative, it typically converts to a
large unsigned number (because unsigned->signed conversion is
+(UINT_MAX+1) %(UINT_MAX+1)), and surprise, you're asking for a
ridiculous malloc size.

Now, reasonable citizens might feel that if your code is using
negative n when calculating a malloc size, you already screwed up, but
gcc has no truck with that. Regardless of what n *actually* is, gcc
*assumes* that n could be any negative number. (n.b. the compiler is
smart enough to detect some range constraints, but not smart enough to
see them all; here, for example, it's not seeing that ESL_GETOPTS is
range checking n>0 in the --cpu option.) If this means that the
unsigned/signed arithmetic works out to a malloc size of larger than
PTRDIFF_MAX (typically 2^63-1, about 9.2 exabytes), it issues the
warning.

The fix is to cast the size_t to a signed integer, and the most
standard-worthy choice of signed integer type is ptrdiff_t. Did this
in seven places where this exact construct occurs.

This fix supersedes an earlier fix (iss #219; commit #094e0d6 7 Jan
2021) which only fixed phmmer.c, and without really understanding the
root of the problem.

(xref SRE:2023/0804-h3-gcc-malloc-warning)