Adds BLOSUM62 from easel/formats to testsuite/
And updates testsuite.sqc paths to use it. This fixes an issue in
infernal where if you configure from a different directory (build dir) without having
already done a top-level configure, you'll get hmmer test failures for the two tests
that use easel/formats/BLOSUM62 because 'easel/formats' doesn't exist as a subdir of
hmmer in the build dir (it does exist in the src dir).