Fixes FM-index code that assumed char is signed.
Fix for issue #318. makehmmerdb.c and fm_alphabet.c were using -1 as
an "unset" flag in `char *inv_alpha` in the FM_METADATA structure.

To reproduce, compile with `gcc -g -Wall -funsigned-char`, which will
give warnings; or build on Power or ARM platform with `./configure
--enable-debugging`. I had tested on Power and ARM platforms
systematically, but only with `./configure` and no warning flags; this
was a mistake. Noted that in the last HMMER release protocol notes.

Changed the unset flag for inv_alpha from -1 to 127. Hacky but
effective. (I considered changing the type to int8_t, but this created
new issues because of char * array allocations and function args.)

Added comments to explain the use of hard-coded 127, with xrefs
to notes.

The code should have probably used ESL_DSQ to begin with, instead of
coming up with its own digital alphabet. Added comment at the top of
fm_alphabet.c that we should look at this if we ever refactor.