Adds --enable-msan to ./configure, to enable memory sanitizer
Following the same addition to Easel's configure script.

The memory sanitizer is only supported on clang and Linux, not gcc or
MacOS. To use:

    ./configure CC=clang --enable-msan

The p7_gmx unit test includes a massive allocation that sanitizers
typically can't handle; added a conditional flag for !eslENABLE_MSAN
that wraps that particular test.